+++ NEWS.md
@@ -1,4 +1,4 @@
-# rstudioapi (development version)
+# rstudioapi 0.19.0
 
 * `getDelegatedAzureToken()` gains an `as` argument. With `as = "AzureToken"`,
   the token is returned as an R6 object compatible with the `AzureToken` class